fixes plot error when no FLEN/TLEN pairs exist
fn125_fl_gt_lt crashed when a species list was provided for which there were no paired observations of flen and tlen. Not detected earlier as this test was previously written to skip in nrows < 10 which didn't makes sense for this test since N=1 should still be flagged as an error and doesn't require a model to identify the error.